api/requirements.txt
fastapi
uvicorn[standard] # Includes websockets and http-tools for performance
python-dotenv